[ot] hw/opentitan: ot_aes: fix IV updates in CFB mode
In AES CFB mode after a block operation the IV registers are updated with
the ciphertext from the previous block operation. See:

- https://opentitan.org/book/hw/ip/aes/doc/theory_of_operation.html#datapath-architecture-and-operation
- https://en.wikipedia.org/wiki/Block_cipher_mode_of_operation#Cipher_feedback_(CFB)

LibTomCrypt does not provide this, so we just read it directly from our
own data. Previously, we were using the updated IV value that LibTomCrypt
provides, which is instead updated to be the output of the block operation
(without the XOR with the plaintext).

This fix allows the aes_functest to pass.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>